adjustments from commentary
