+++ app/data/schemas.server.ts
@@ -142,436 +142,3 @@ export const BlessingSchema = z
       baseName,
     }
   })
-
-  
-export const AuthSchema = z.object({
-  AccessToken: z.string(),
-  ExpiresIn: z.number(),
-  RefreshToken: z.string(),
-  Sub: z.string(),
-})
-export const MissionBoardSchema = z.object({
-  missions: z.array(
-    z.object({
-      id: z.string(),
-      map: z.string(),
-      circumstance: z.string(),
-      flags: z.unknown(),
-      credits: z.number(),
-      xp: z.number(),
-      extraRewards: z.unknown(),
-      challenge: z.number(),
-      resistance: z.number(),
-      start: z.string(),
-      expiry: z.string(),
-      requiredLevel: z.number(),
-      missionGiver: z.string(),
-      displayIndex: z.number(),
-    })
-  ),
-  refreshAt: z.string(),
-  _links: z.unknown(),
-})
-export const LatenciesSchema = z.object({
-  regions: z.array(
-    z.object({
-      region: z.string(),
-      httpLatencyUrl: z.string(),
-    })
-  ),
-})
-
-export const AccountSummarySchema = z.object({
-  summary: z.object({
-    sub: z.string(),
-    name: z.string(),
-    characters: z.array(
-      z.object({
-        id: z.string(),
-        name: z.string(),
-        inventory: z.record(z.string()),
-        breed: z.string(),
-        archetype: z.string(),
-        gender: z.string(),
-        lore: z.unknown(),
-        selectedVoice: z.string().optional(),
-        abilities: z.unknown(),
-        career: z.object({
-          specialization: z.string(),
-          talents: z.array(z.string()).optional(),
-        }),
-        narrative: z.unknown(),
-        personal: z.unknown(),
-        prison_number: z.string(),
-      })
-    ),
-    data: z.array(z.unknown()),
-    currencies: z.array(z.unknown()),
-    statistics: z.array(z.unknown()),
-    consumables: z.array(z.unknown()),
-    activeEffects: z.array(z.unknown()),
-    progression: z.object({
-      type: z.literal("account"),
-      id: z.string(),
-      currentLevel: z.number(),
-      currentXp: z.number(),
-      currentXpInLevel: z.number(),
-      neededXpForNextLevel: z.number(),
-      unclaimedRewards: z.array(z.unknown()),
-      eligibleForLevel: z.boolean(),
-    }),
-  }),
-})
-export const AccountGearSchema = z.object({
-  gearList: z.record(
-    z.object({
-      slots: z.array(z.string()),
-      masterDataInstance: z.object({
-        id: z.string(),
-        overrides: z
-          .object({
-            ver: z.number().optional(),
-            traits: z
-              .array(
-                z.object({
-                  value: z.number().optional(),
-                  id: z.string(),
-                  rarity: z.number(),
-                })
-              )
-              .optional(),
-            perks: z
-              .array(
-                z.object({
-                  value: z.number().optional(),
-                  id: z.string(),
-                  rarity: z.number(),
-                })
-              )
-              .optional(),
-            characterLevel: z.number().optional(),
-            baseItemLevel: z.number().optional(),
-            baseStats: z
-              .array(z.object({ name: z.string(), value: z.number() }))
-              .optional(),
-            rarity: z.number().optional(),
-            itemLevel: z.number().optional(),
-          })
-          .optional(),
-      }),
-      characterId: z.string().optional(),
-    })
-  ),
-})
-
-export const JoinSchema = z.object({
-  queuePosition: z.number(),
-  queueTicket: z.string(),
-  retrySuggestion: z.number(),
-})
-
-
-export const CharactersSchema = z.object({
-  _links: z.object({ self: z.object({ href: z.string() }) }),
-  characters: z.array(
-    z.union([
-      z.object({
-        _links: z.object({
-          self: z.object({ href: z.string() }),
-          data: z.object({ href: z.string(), templated: z.boolean() }),
-          statistics: z.object({ href: z.string(), templated: z.boolean() }),
-          history: z.object({ href: z.string(), templated: z.boolean() }),
-          inventory: z.object({ href: z.string() }),
-          gear: z.object({ href: z.string() })
-        }),
-        id: z.string(),
-        name: z.string(),
-        inventory: z.object({
-          slot_animation_emote_1: z.string(),
-          slot_animation_emote_2: z.string(),
-          slot_animation_emote_3: z.string(),
-          slot_animation_emote_4: z.string(),
-          slot_animation_emote_5: z.string(),
-          slot_attachment_2: z.string(),
-          slot_attachment_3: z.string(),
-          slot_attachment_1: z.string(),
-          slot_body_legs: z.string(),
-          slot_body_face_scar: z.string(),
-          slot_secondary: z.string(),
-          slot_body_hair: z.string(),
-          slot_body_arms: z.string(),
-          slot_body_skin_color: z.string(),
-          slot_portrait_frame: z.string(),
-          slot_body_face: z.string(),
-          slot_animation_end_of_round: z.string(),
-          slot_gear_lowerbody: z.string(),
-          slot_body_eye_color: z.string(),
-          slot_insignia: z.string(),
-          slot_gear_extra_cosmetic: z.string(),
-          slot_body_torso: z.string(),
-          slot_body_face_tattoo: z.string(),
-          slot_body_hair_color: z.string(),
-          slot_body_tattoo: z.string(),
-          slot_primary: z.string(),
-          slot_body_face_hair: z.string(),
-          slot_gear_upperbody: z.string(),
-          slot_gear_head: z.string()
-        }),
-        breed: z.string(),
-        archetype: z.string(),
-        gender: z.string(),
-        lore: z.object({
-          backstory: z.object({
-            crime: z.string(),
-            formative_event: z.string(),
-            growing_up: z.string(),
-            childhood: z.string(),
-            personality: z.string(),
-            planet: z.string()
-          })
-        }),
-        selected_voice: z.string(),
-        abilities: z.object({
-          grenade_ability: z.string(),
-          combat_ability: z.string()
-        }),
-        career: z.object({
-          specialization: z.string(),
-          talents: z.array(z.string())
-        }),
-        narrative: z.object({
-          events: z.object({
-            onboarding_step_mission_board_introduction: z.boolean(),
-            level_unlock_crafting_station_visited: z.boolean(),
-            level_unlock_credits_store_visited: z.boolean(),
-            level_unlock_premium_store_visited: z.boolean(),
-            mission_board: z.boolean(),
-            level_unlock_barber_visited: z.boolean(),
-            onboarding_step_chapel_video_viewed: z.boolean(),
-            onboarding_step_chapel_cutscene_played: z.boolean(),
-            level_unlock_contract_store_visited: z.boolean()
-          }),
-          stories: z.object({
-            onboarding: z.number(),
-            level_unlock_popups: z.number(),
-            path_of_trust: z.number()
-          })
-        }),
-        personal: z.object({ character_height: z.number() }),
-        prison_number: z.string(),
-        memberships: z.array(z.unknown())
-      }),
-      z.object({
-        _links: z.object({
-          self: z.object({ href: z.string() }),
-          data: z.object({ href: z.string(), templated: z.boolean() }),
-          statistics: z.object({ href: z.string(), templated: z.boolean() }),
-          history: z.object({ href: z.string(), templated: z.boolean() }),
-          inventory: z.object({ href: z.string() }),
-          gear: z.object({ href: z.string() })
-        }),
-        id: z.string(),
-        name: z.string(),
-        inventory: z.object({
-          slot_attachment_2: z.string(),
-          slot_attachment_1: z.string(),
-          slot_body_legs: z.string(),
-          slot_body_face_scar: z.string(),
-          slot_secondary: z.string(),
-          slot_body_hair: z.string(),
-          slot_body_arms: z.string(),
-          slot_body_skin_color: z.string(),
-          slot_portrait_frame: z.string(),
-          slot_body_face: z.string(),
-          slot_animation_end_of_round: z.string(),
-          slot_gear_lowerbody: z.string(),
-          slot_body_eye_color: z.string(),
-          slot_insignia: z.string(),
-          slot_gear_extra_cosmetic: z.string(),
-          slot_body_torso: z.string(),
-          slot_body_face_tattoo: z.string(),
-          slot_body_hair_color: z.string(),
-          slot_body_tattoo: z.string(),
-          slot_primary: z.string(),
-          slot_body_face_hair: z.string(),
-          slot_gear_upperbody: z.string(),
-          slot_gear_head: z.string()
-        }),
-        breed: z.string(),
-        archetype: z.string(),
-        gender: z.string(),
-        lore: z.object({
-          backstory: z.object({
-            crime: z.string(),
-            formative_event: z.string(),
-            growing_up: z.string(),
-            childhood: z.string(),
-            personality: z.string(),
-            planet: z.string()
-          })
-        }),
-        selected_voice: z.string(),
-        abilities: z.object({
-          grenade_ability: z.string(),
-          combat_ability: z.string()
-        }),
-        career: z.object({
-          specialization: z.string(),
-          talents: z.array(z.string())
-        }),
-        narrative: z.object({
-          events: z.object({
-            level_unlock_crafting_station_visited: z.boolean(),
-            onboarding_step_mission_board_introduction: z.boolean(),
-            level_unlock_credits_store_visited: z.boolean(),
-            level_unlock_premium_store_visited: z.boolean(),
-            mission_board: z.boolean(),
-            level_unlock_barber_visited: z.boolean(),
-            onboarding_step_chapel_video_viewed: z.boolean(),
-            onboarding_step_chapel_cutscene_played: z.boolean(),
-            level_unlock_contract_store_visited: z.boolean()
-          }),
-          stories: z.object({
-            onboarding: z.number(),
-            level_unlock_popups: z.number(),
-            path_of_trust: z.number()
-          })
-        }),
-        personal: z.object({ character_height: z.number() }),
-        prison_number: z.string(),
-        memberships: z.array(z.unknown())
-      }),
-      z.object({
-        _links: z.object({
-          self: z.object({ href: z.string() }),
-          data: z.object({ href: z.string(), templated: z.boolean() }),
-          statistics: z.object({ href: z.string(), templated: z.boolean() }),
-          history: z.object({ href: z.string(), templated: z.boolean() }),
-          inventory: z.object({ href: z.string() }),
-          gear: z.object({ href: z.string() })
-        }),
-        id: z.string(),
-        name: z.string(),
-        inventory: z.object({
-          slot_body_legs: z.string(),
-          slot_body_face_scar: z.string(),
-          slot_secondary: z.string(),
-          slot_body_hair: z.string(),
-          slot_body_arms: z.string(),
-          slot_body_skin_color: z.string(),
-          slot_portrait_frame: z.string(),
-          slot_body_face: z.string(),
-          slot_animation_end_of_round: z.string(),
-          slot_gear_lowerbody: z.string(),
-          slot_body_eye_color: z.string(),
-          slot_insignia: z.string(),
-          slot_gear_extra_cosmetic: z.string(),
-          slot_body_torso: z.string(),
-          slot_body_face_tattoo: z.string(),
-          slot_body_hair_color: z.string(),
-          slot_body_tattoo: z.string(),
-          slot_primary: z.string(),
-          slot_body_face_hair: z.string(),
-          slot_gear_upperbody: z.string(),
-          slot_gear_head: z.string()
-        }),
-        breed: z.string(),
-        archetype: z.string(),
-        gender: z.string(),
-        lore: z.object({
-          backstory: z.object({
-            crime: z.string(),
-            formative_event: z.string(),
-            growing_up: z.string(),
-            childhood: z.string(),
-            personality: z.string(),
-            planet: z.string()
-          })
-        }),
-        selected_voice: z.string(),
-        abilities: z.object({
-          grenade_ability: z.string(),
-          combat_ability: z.string()
-        }),
-        career: z.object({ specialization: z.string() }),
-        narrative: z.object({
-          events: z.object({
-            onboarding_step_chapel_video_viewed: z.boolean(),
-            onboarding_step_chapel_cutscene_played: z.boolean()
-          }),
-          stories: z.object({ onboarding: z.number() })
-        }),
-        personal: z.object({ character_height: z.number() }),
-        prison_number: z.string(),
-        memberships: z.array(z.unknown())
-      })
-    ])
-  )
-})
-
-export const ShopSchema = z.object({
-  _links: z.object({
-    self: z.object({ href: z.string() }),
-    config: z.object({ href: z.string() })
-  }),
-  catalog: z.object({
-    id: z.string(),
-    name: z.string(),
-    generation: z.number(),
-    layoutRef: z.string(),
-    validFrom: z.string(),
-    validTo: z.string()
-  }),
-  name: z.string(),
-  public: z.array(z.unknown()),
-  personal: z.array(
-        z.object({
-        offerId: z.string(),
-        sku: z.object({
-          id: z.string(),
-          displayPriority: z.number(),
-          internalName: z.string(),
-          name: z.string(),
-          description: z.string(),
-          category: z.string(),
-          assetId: z.string(),
-          tags: z.array(z.unknown()),
-          dlcReq: z.array(z.unknown())
-        }),
-        entitlement: z.object({
-          id: z.string(),
-          limit: z.number(),
-          type: z.string()
-        }),
-        price: z.object({
-          amount: z.object({ amount: z.number(), type: z.string() }),
-          id: z.string(),
-          priority: z.number(),
-          priceFormula: z.string()
-        }),
-        state: z.string(),
-        description: z.object({
-          id: z.string(),
-          gearId: z.string(),
-          rotation: z.string(),
-          type: z.string(),
-          properties: z.object({}),
-          overrides: z.object({
-            ver: z.number(),
-            rarity: z.number(),
-            characterLevel: z.number(),
-            itemLevel: z.number(),
-            baseItemLevel: z.number(),
-            traits: z.array(z.unknown()),
-            perks: z.array(z.object({id: z.string(), rarity: z.number()})).optional(),
-            base_stats: z.array(
-              z.object({ name: z.string(), value: z.number() })
-            ).optional()
-          })
-        }),
-        media: z.array(z.unknown())
-      }).optional(),    
-  ),
-  rerollsThisRotation: z.number(),
-  currentRotationEnd: z.string()
-})
