+++ app/root.tsx
@@ -16,9 +16,8 @@ import { authenticator } from "~/services/auth.server"
 // Sets up Cron singletons to perform timed jobs on the server
 import { initJobs, getTunnel } from "~/jobs/index.server"
 initJobs && initJobs()
-if( process.env.NODE_ENV !== "production") {	
-	getTunnel && getTunnel()
-}
+getTunnel && getTunnel()
+
 
 export const links: LinksFunction = () => [
 	{ rel: "stylesheet", href: tailwind },
